#135 refactor: parsing resource group from id function as a library
