+++ src/spaceone/inventory/manager/disk_manager.py
@@ -98,11 +98,6 @@ class DiskManager(AzureManager):
         return disks
 
     @staticmethod
-    def get_resource_group_from_id(disk_id):
-        resource_group = disk_id.split('/')[4].lower()
-        return resource_group
-
-    @staticmethod
     def get_attached_vm_name_from_managed_by(managed_by):
         attached_vm_name = managed_by.split('/')[8]
         return attached_vm_name
